lib/rigging/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "marin-rigging"
version = "0.99"
requires-python = ">=3.11,<3.13"
dependencies = [
    "fsspec>=2024.0.0",
    "gcsfs>=2024.0.0",
    "PyJWT>=2.12.0",
    "s3fs>=2024.0.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.3.2",
    "pytest-timeout",
]

[tool.hatch.build.targets.wheel]
packages = ["src/rigging"]

[tool.pytest.ini_options]
timeout = 10
